requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
django-markdown==0.8.4
django-taggit==3.0.0
importlib-metadata==4.12.0
Markdown==3.3.7
psycopg2-binary==2.9.3
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
zipp==3.8.0
